% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_fig3.R
\name{dataprep_fig3}
\alias{dataprep_fig3}
\title{Dataprep Figure 3}
\usage{
dataprep_fig3(postdir)
}
\arguments{
\item{postdir}{A filepath specifying where the posteior samples that are generated
from the combine_posteriors function are saved. are saved. If outputs have not been
moved, this will be in a directory "/MajorGroups".}
}
\description{
This takes the posterior values of the group level indicators generated within the
\code{generage_fig1} function, calculates the quantiles and plots the group level quantiles presented in figure 3.
\code{dataprep_fig3} is called within  \code{generage_fig3} by setting \code{dataprep = TRUE}.
}
\examples{
\dontrun{

# Run generate_fig3 function to produce quantile plots
# datadir should be the filepath of where the posterior indicator values are saved.
# generate_fig3(postdir = paste0(getwd(), "/MajorGroups/geomeans"))

}
}
