% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fig1.r
\name{generate_fig1}
\alias{generate_fig1}
\title{Generating Figure 1}
\usage{
generate_fig1(postdir, status = TRUE)
}
\arguments{
\item{postdir}{A filepath specifying where the posterior combinations are saved.}

\item{status}{Logical.  If `TRUE`, group name will be printed to the console
as means and indicator values are being estimated. Default is `TRUE`.}
}
\description{
This takes the major group level posterior combinations generated using the
\code{combine_posteriors} function and generates figure 1 presented within the paper.
It also generates the posteriors of average change across species for each group that
are also used in the estimation of major group level trends presented within the text
of the associated paper and for the generation of figure 3.
}
\examples{
\dontrun{

# Run generate_fig1 function
# postdir should be the filepath of where the 4 major group level posteriors
# combinations are saved.
generate_fig1(postdir = paste0(getwd(), "/MajorGroups"))

}
}
\references{
Outhwaite et al (in prep) Complexity of biodiversity change revealed through long-term trends of invertebrates, bryophytes and lichens.

Outhwaite, C. L., Powney, G. D., August, T. A., Chandler, R. E., Rorke, S., Pescott, O., … Isaac, N. J. B. (2019). Annual estimates of
 occupancy for bryophytes, lichens and invertebrates in the UK (1970-2015).
 NERC Environmental Information Data Centre. https://doi.org/10.5285/0ec7e549-57d4-4e2d-b2d3-2199e1578d84
}
\keyword{distribution,}
\keyword{occupancy}
\keyword{species,}
\keyword{trends,}
