% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fig2.r
\name{generate_fig2}
\alias{generate_fig2}
\title{Generate Figure 2}
\usage{
generate_fig2(datadir, save_plot = TRUE)
}
\arguments{
\item{datadir}{A filepath specifying where the posteior indicator values are saved.
If outputs have not been moved, this will be in a directory "/MajorGroups/geomeans".}

\item{save_plot}{Logical. If `TRUE` plot will be saved in `datadir`}
}
\description{
This takes the posterior values of the group level indicators generated within the
\code{generage_fig1} function and generates the boxplots presented in figure 2 of the associated paper.
}
\examples{
\dontrun{

# Run generate_fig2 function to produce boxplot.
# datadir should be the filepath of where the posterior indicator values are saved.
generate_fig2(postdir = paste0(getwd(), "/MajorGroups/geomeans"))

}
}
\references{
Outhwaite et al (in prep) Complexity of biodiversity change revealed through long-term trends of invertebrates, bryophytes and lichens.

Outhwaite, C. L., Powney, G. D., August, T. A., Chandler, R. E., Rorke, S., Pescott, O., … Isaac, N. J. B. (2019). Annual estimates of
 occupancy for bryophytes, lichens and invertebrates in the UK (1970-2015).
 NERC Environmental Information Data Centre. https://doi.org/10.5285/0ec7e549-57d4-4e2d-b2d3-2199e1578d84
}
\keyword{distribution,}
\keyword{occupancy}
\keyword{species,}
\keyword{trends,}
