% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fig3.r
\name{generate_fig3}
\alias{generate_fig3}
\title{Generate Figure 3}
\usage{
generate_fig3(postdir, save_plot = TRUE, interval = 95,
  dataprep = TRUE)
}
\arguments{
\item{postdir}{A filepath specifying where the posteior samples that are generated
from the combine_posteriors function are saved. are saved. If outputs have not been
moved, this will be in a directory "/MajorGroups".}

\item{save_plot}{Logical.  If `TRUE`, the plot will be saved as a PDF file
within the `postdir`. Default is `TRUE`.}

\item{dataprep}{Logical. Default is `TRUE`, which calls \code{dataprep_fig3} to prepare the data. 
Having run \code{generage_fig3} once with \code{dataprep = TRUE}, the figure can be redrawn quickly with \code{dataprep = FALSE}, e.g. using different values for `interval`.}
}
\description{
This takes the posterior values of the group level indicators generated within the
\code{generage_fig1} function, calculates the quantiles and plots the group level quantiles presented in figure 3.
}
\examples{
\dontrun{

# Run generate_fig3 function to produce quantile plots
# datadir should be the filepath of where the posterior indicator values are saved.
# generate_fig3(postdir = paste0(getwd(), "/MajorGroups/geomeans"))

}
}
\references{
Outhwaite et al (in prep) Complexity of biodiversity change revealed through long-term trends of invertebrates, bryophytes and lichens.

Outhwaite, C. L., Powney, G. D., August, T. A., Chandler, R. E., Rorke, S., Pescott, O., … Isaac, N. J. B. (2019). Annual estimates of
 occupancy for bryophytes, lichens and invertebrates in the UK (1970-2015).
 NERC Environmental Information Data Centre. https://doi.org/10.5285/0ec7e549-57d4-4e2d-b2d3-2199e1578d84
}
\keyword{distribution,}
\keyword{occupancy}
\keyword{species,}
\keyword{trends,}
