% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fig4.r
\name{generate_fig4}
\alias{generate_fig4}
\title{Generating Figure 4}
\usage{
generate_fig4(postdir, save_plot = TRUE, interval = 95)
}
\arguments{
\item{postdir}{A filepath specifying where the posterior combinations are saved.}

\item{save_plot}{Logical.  If `TRUE`, the plot will be saved as a PDF file
within the `postdir`. Default is `TRUE`.}
}
\description{
This takes the taxa level posterior combinations generated using the
\code{combine_posteriors} function and generates figure 4 presented within the paper.
}
\examples{
\dontrun{

# Run generate_fig4 function
# postdir should be the filepath of where the 4 major group level posteriors
# combinationss are saved.
generate_fig4(postdir = paste0(getwd(), "/Taxa"))

}
}
\references{
Outhwaite et al (in prep) Complexity of biodiversity change revealed through long-term trends of invertebrates, bryophytes and lichens.

Outhwaite, C. L., Powney, G. D., August, T. A., Chandler, R. E., Rorke, S., Pescott, O., … Isaac, N. J. B. (2019). Annual estimates of
 occupancy for bryophytes, lichens and invertebrates in the UK (1970-2015).
 NERC Environmental Information Data Centre. https://doi.org/10.5285/0ec7e549-57d4-4e2d-b2d3-2199e1578d84
}
\keyword{distribution,}
\keyword{occupancy}
\keyword{species,}
\keyword{trends,}
