% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_suppfig2.r
\name{generate_suppfig2}
\alias{generate_suppfig2}
\title{Generating Supplementary Figure 2}
\usage{
generate_suppfig2(postdir, sp_trends, status = TRUE, save_plot = TRUE)
}
\arguments{
\item{postdir}{A filepath specifying where the posterior combinations are saved.}

\item{sp_trends}{A dataframe, downloaded from the repository which details the species
level trends in occupancy and the years for which data were avaialble for each species.}

\item{status}{Logical. If `TRUE` the progress through each group and number of records subset will be printed to the console.}

\item{save_plot}{Logical.  If `TRUE`, the plot will be saved as a PDF file
within the `outdir`. Default is `TRUE`.}
}
\description{
This generates version of figure 1 but with subsets of species based on the number of records available for the occupancy estimation.
}
\examples{
\dontrun{

# Run generate_fig1 function
# postdir should be the filepath of where the 4 major group level posteriors
# combinationss are saved.
generate_fig2supp(postdir = paste0(getwd(), "/MajorGroups"),
sp_trends = read.csv(paste0(datadir, "/Species_Trends.csv")),
outdir = paste0(getwd(), "/MajorGroups"))

}
}
\references{
Outhwaite et al (in prep) Complexity of biodiversity change revealed through long-term trends of invertebrates, bryophytes and lichens.

Outhwaite, C. L., Powney, G. D., August, T. A., Chandler, R. E., Rorke, S., Pescott, O., … Isaac, N. J. B. (2019). Annual estimates of
 occupancy for bryophytes, lichens and invertebrates in the UK (1970-2015).
 NERC Environmental Information Data Centre. https://doi.org/10.5285/0ec7e549-57d4-4e2d-b2d3-2199e1578d84
}
\keyword{distribution,}
\keyword{occupancy}
\keyword{species,}
\keyword{trends,}
