% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r, R/major_groups.r
\docType{data}
\name{major_groups}
\alias{major_groups}
\title{Details of major group of each species.}
\format{A data frame with 5293 rows and 3 variables:
\describe{
  \item{Group}{Taxa name, 32 levels}
  \item{Species}{Species name}
  \item{Major_group}{Major group, e.g. FREHWATER_SPECIES, 4 levels}
  ...
}}
\usage{
major_groups

major_groups
}
\description{
A dataset containing each species name and the associated taxa and major_group.
This dataset is necessary for generating the group level analyses.

A dataset containing information on which major group (e.g. FRESHWATER_SPECIES
 or "LOWER_PLANTS") each species has been assigned to for the analysis carried
 out in the paper by Outhwaite et al.
}
\keyword{datasets}
